A bit of cleanup
Cleans up bots.rb.
Moves method into separate file (because it might get longer)
Only takes up a single method now (instead of three): #config
